requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Database - Async First
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
supabase==2.5.1

# Redis & Event Streaming
redis==5.0.7
hiredis==2.3.2

# API Clients
httpx==0.27.0
gspread==6.1.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1

# OpenAI - Latest with Responses API
openai>=1.35.0

# Slack
slack-bolt==1.19.0
slack-sdk==3.31.0
aiohttp==3.9.5

# Background Tasks
celery==5.4.0
flower==2.0.1
celery-redbeat==2.2.0

# Monitoring & Logging
structlog==24.2.0
python-json-logger==2.0.7
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-sqlalchemy==0.46b0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
pendulum==3.0.0
orjson==3.10.5
python-dateutil==2.9.0

# Data Processing
pandas==2.2.2
numpy==1.26.4
scipy==1.14.0

# Geospatial
shapely==2.0.4
geopy==2.4.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-httpx==0.30.0
pytest-mock==3.14.0
respx==0.21.1

# Development
black==24.4.2
ruff==0.5.0
mypy==1.10.1
pre-commit==3.7.1
ipython==8.25.0

# Type stubs
types-redis==4.6.0
types-requests==2.32.0
pandas-stubs==2.2.2
